moves methods from Rules to Board and delete Rules and RulesTest
